Nova tanda de commits pel process de signature module
git-svn-id: https://svn.code.sf.net/p/portafib/code-0/portafib-core/branches/portafib-1.1@185 1a83b7bb-df65-4973-8540-7514423c1e25

+++ back/src/main/java/es/caib/portafib/back/controller/webdb/PeticioDeFirmaController.java
@@ -992,7 +992,7 @@ public java.lang.Long stringToPK(String value) {
     Where _w = null;
     if (!_groupByItemsMap.containsKey(ALGORISMEDEFIRMAID)) {
       // OBTENIR TOTES LES CLAUS (PK) i despres només cercar referències d'aquestes PK
-      java.util.Set<java.lang.Long> _pkList = new java.util.HashSet<java.lang.Long>();
+      java.util.Set<java.lang.Integer> _pkList = new java.util.HashSet<java.lang.Integer>();
       for (PeticioDeFirma _item : list) {
         _pkList.add(_item.getAlgorismeDeFirmaID());
         }
