+++ back/src/main/java/es/caib/portafib/back/form/webdb/PeticioDeFirmaFilterForm.java
@@ -219,24 +219,24 @@ public class PeticioDeFirmaFilterForm extends PortaFIBBaseFilterForm implements 
   }
 
 
-  private java.lang.Long algorismeDeFirmaIDDesde;
+  private java.lang.Integer algorismeDeFirmaIDDesde;
 
-  public java.lang.Long getAlgorismeDeFirmaIDDesde() {
+  public java.lang.Integer getAlgorismeDeFirmaIDDesde() {
     return this.algorismeDeFirmaIDDesde;
   }
 
-  public void setAlgorismeDeFirmaIDDesde(java.lang.Long algorismeDeFirmaIDDesde) {
+  public void setAlgorismeDeFirmaIDDesde(java.lang.Integer algorismeDeFirmaIDDesde) {
     this.algorismeDeFirmaIDDesde = algorismeDeFirmaIDDesde;
   }
 
 
-  private java.lang.Long algorismeDeFirmaIDFins;
+  private java.lang.Integer algorismeDeFirmaIDFins;
 
-  public java.lang.Long getAlgorismeDeFirmaIDFins() {
+  public java.lang.Integer getAlgorismeDeFirmaIDFins() {
     return this.algorismeDeFirmaIDFins;
   }
 
-  public void setAlgorismeDeFirmaIDFins(java.lang.Long algorismeDeFirmaIDFins) {
+  public void setAlgorismeDeFirmaIDFins(java.lang.Integer algorismeDeFirmaIDFins) {
     this.algorismeDeFirmaIDFins = algorismeDeFirmaIDFins;
   }
