+++ jpa/src/main/java/es/caib/portafib/jpa/validator/AlgorismeDeFirmaValidator.java
@@ -72,7 +72,7 @@ public class AlgorismeDeFirmaValidator<T> implements AlgorismeDeFirmaFields {
 
       // Check Unique - PK no AutoIncrement amb UNA SOLA PK 
       if (__vr.getFieldErrorCount(ALGORISMEDEFIRMAID) == 0) {
-        java.lang.Long __algorismedefirmaid = (java.lang.Long)__vr.getFieldValue(__target__,ALGORISMEDEFIRMAID);
+        java.lang.Integer __algorismedefirmaid = (java.lang.Integer)__vr.getFieldValue(__target__,ALGORISMEDEFIRMAID);
         Long __count_ = null;
         try { __count_ = __algorismeDeFirmaManager.count(org.fundaciobit.genapp.common.query.Where.AND(ALGORISMEDEFIRMAID.equal(__algorismedefirmaid))); } catch(org.fundaciobit.genapp.common.i18n.I18NException e) { e.printStackTrace(); };
         if (__count_ == null || __count_ != 0) {        
@@ -90,7 +90,7 @@ public class AlgorismeDeFirmaValidator<T> implements AlgorismeDeFirmaFields {
       // Check Unique - no PK
       if (__vr.getFieldErrorCount(NOM) == 0 && __vr.getFieldErrorCount(ALGORISMEDEFIRMAID) == 0) {
         java.lang.String __nom = (java.lang.String)__vr.getFieldValue(__target__,NOM);
-        java.lang.Long __algorismedefirmaid = (java.lang.Long)__vr.getFieldValue(__target__,ALGORISMEDEFIRMAID);
+        java.lang.Integer __algorismedefirmaid = (java.lang.Integer)__vr.getFieldValue(__target__,ALGORISMEDEFIRMAID);
         Long __count_ = null;
         try { __count_ = __algorismeDeFirmaManager.count(org.fundaciobit.genapp.common.query.Where.AND(NOM.equal(__nom), ALGORISMEDEFIRMAID.notEqual(__algorismedefirmaid))); } catch(org.fundaciobit.genapp.common.i18n.I18NException e) { e.printStackTrace(); };
         if (__count_ == null || __count_ != 0) {
